Refactor mosaic block renderers for block documents
