+++ packages/mosaic/src/data-table-explorer/block-document/DataTableBlockHeader.tsx
